Updated the UserMeetingSettings controller & service.
Removed custom error handling logic previously implemented in the controller. Now its handled by the default error handler in app.js file & the error is invoked from the service.

Updated the 'handleConnectCalendarCallback' method in the service to stop creating duplicate calendars. And when connecting a new calendar and making it primary, the method makes the other existing calendars in the UserMeetingSettings record non-primary.

Additionally, updated the 'deleteUserCalendar' method in the service to give better error messages and to make another existing calendar as primary when deleting a primary calendar. Also, this method now envokes errors and lets the default error handler instead of the controller to handle them.