Updated the Nylas Calendar schema in UserMeetingSettings model.
Now the calendar schema has one new field - 'isDeleted'. Instead of deleting the calendars in the 'nylasCalendars' array in UserMeetingSettings through the 'DELETE' endpoint, calendars are now stored and the property 'isDeleted' is used to indicate whether the calendar is deleted or not. The UserMeetingSettingsService was updated to accommodate for this change. These changes are as follows:

1. When using the service method 'getUserMeetingSettingsByUserId' to retrieve UserMeetingSettings, the calendars with 'isDeleted' property equal to true are not returned;
2. When creating new UserMeetingSettings record, the associated calendar now includes the new property 'isDeleted' with default value set to false;
3. When connecting a new calendar, if the calendar used to exist before (searched by the calendar id), then instead of creating another one, the previous record is updated with new details and the property 'isDeleted' is set to false;
4. When deleting a calendar, another calendar is only set to 'isPrimary' equal to true if it is not deleted, i.e., 'isDeleted' is false;

Additionally, the Nylas calendar schema now sets 'isDeleted' property in the DB as a non-null boolean.