Added new field 'email' in NylasCalendars column in UserMeetingSettings record.
In NylasService, creating a virtual calendar returns the result, but extends it with new email field - populated with current user's email. The method 'getAccessToken' is also modified to include the server sent email field in the parsed result.

And in UserMeetingSettingsService, method 'stripUnwantedData' was accidentally not stripping the data correctly, but its fixed now. Also, the method 'handleConnectCalendarCallback' has been updated to include the email field when creating/updating calendars in UserMeetingSettings record.