Implemented usage of metadata from event details in Nylas global webhook.
When a new event is created the dedicated create event handler webhook adds identification data in metadata of the event via a PUT request. Then this metadata is used in the global webhook to identify Interviews. Previously, the required data was parsed from string values which was not too reliable.