Updated Nylas global & specific webooks to populate & utilize new Interview fields.
In the webhook for created events, the event properties 'calendar_event_id' & 'edit_hash' are saved in the fields 'nylasEventId' & 'nylasEventEditHash' respectively in the associated Interview record.
And in the global Nylas webhook endpoint, an interview's status is only updated as cancelled if the event id matches with the 'nylasEventId' property of the associated Interview record & if the status is not already cancelled.